+++ sdk/storage/azure-storage-blob-cryptography/src/main/java/com/azure/storage/blob/specialized/cryptography/EncryptedBlobClient.java
@@ -4,19 +4,26 @@
 package com.azure.storage.blob.specialized.cryptography;
 
 import com.azure.core.annotation.ServiceClient;
+import com.azure.core.util.Context;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.storage.blob.BlobClient;
 import com.azure.storage.blob.BlobClientBuilder;
+import com.azure.storage.blob.implementation.util.ModelHelper;
 import com.azure.storage.blob.models.AccessTier;
 import com.azure.storage.blob.models.BlobRequestConditions;
 import com.azure.storage.blob.models.BlobHttpHeaders;
 import com.azure.storage.blob.models.BlobStorageException;
 import com.azure.storage.blob.models.ParallelTransferOptions;
+import com.azure.storage.blob.specialized.AppendBlobClient;
 import com.azure.storage.blob.specialized.BlobOutputStream;
+import com.azure.storage.blob.specialized.BlockBlobClient;
+import com.azure.storage.blob.specialized.PageBlobClient;
 import com.azure.storage.common.implementation.Constants;
 import com.azure.storage.common.implementation.StorageImplUtils;
 import reactor.core.publisher.Mono;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.io.UncheckedIOException;
 import java.time.Duration;
 import java.util.Map;
@@ -110,6 +117,31 @@ public class EncryptedBlobClient extends BlobClient {
     }
 
     /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void uploadWithResponse(InputStream data, long length, ParallelTransferOptions parallelTransferOptions,
+        BlobHttpHeaders headers, Map<String, String> metadata, AccessTier tier, BlobRequestConditions requestConditions,
+        Duration timeout, Context context) {
+        final ParallelTransferOptions validatedParallelTransferOptions =
+            ModelHelper.populateAndApplyDefaults(parallelTransferOptions);
+
+        // BlobOutputStream passes to buffered upload, which has logic for switching based on stream size
+        BlobOutputStream blobOutputStream = BlobOutputStream.blockBlobOutputStream(encryptedBlobAsyncClient,
+            validatedParallelTransferOptions, headers, metadata, tier, requestConditions);
+        try {
+            StorageImplUtils.copyToOutputStream(data, length, blobOutputStream);
+            blobOutputStream.close();
+        } catch (IOException e) {
+            if (e.getCause() instanceof BlobStorageException) {
+                throw logger.logExceptionAsError((BlobStorageException) e.getCause());
+            } else {
+                throw logger.logExceptionAsError(new UncheckedIOException(e));
+            }
+        }
+    }
+
+    /**
      * Creates a new block blob, or updates the content of an existing block blob.
      *
      * <p><strong>Code Samples</strong></p>
@@ -118,6 +150,7 @@ public class EncryptedBlobClient extends BlobClient {
      *
      * @param filePath Path of the file to upload
      */
+    @Override
     public void uploadFromFile(String filePath) {
         uploadFromFile(filePath, false);
     }
@@ -132,6 +165,7 @@ public class EncryptedBlobClient extends BlobClient {
      * @param filePath Path of the file to upload
      * @param overwrite Whether or not to overwrite should data already exist on the blob
      */
+    @Override
     public void uploadFromFile(String filePath, boolean overwrite) {
         if (!overwrite && exists()) {
             throw logger.logExceptionAsError(new IllegalArgumentException(Constants.BLOB_ALREADY_EXISTS));
@@ -156,6 +190,7 @@ public class EncryptedBlobClient extends BlobClient {
      * @param timeout An optional timeout value beyond which a {@link RuntimeException} will be raised.
      * @throws UncheckedIOException If an I/O error occurs
      */
+    @Override
     public void uploadFromFile(String filePath, ParallelTransferOptions parallelTransferOptions,
         BlobHttpHeaders headers, Map<String, String> metadata, AccessTier tier, BlobRequestConditions requestConditions,
         Duration timeout) throws UncheckedIOException {
@@ -167,6 +202,33 @@ public class EncryptedBlobClient extends BlobClient {
         } catch (UncheckedIOException e) {
             throw logger.logExceptionAsError(e);
         }
+    }
+
+    /**
+     * Unsupported.
+     */
+    @Override
+    public AppendBlobClient getAppendBlobClient() {
+        throw logger.logExceptionAsError(new UnsupportedOperationException("Cannot get an encrypted client as an append"
+            + " blob client"));
+    }
+
+    /**
+     * Unsupported.
+     */
+    @Override
+    public BlockBlobClient getBlockBlobClient() {
+        throw logger.logExceptionAsError(new UnsupportedOperationException("Cannot get an encrypted client as a block"
+            + " blob client"));
+    }
+
+    /**
+     * Unsupported.
+     */
+    @Override
+    public PageBlobClient getPageBlobClient() {
+        throw logger.logExceptionAsError(new UnsupportedOperationException("Cannot get an encrypted client as an page"
+            + " blob client"));
     }
 
 }