+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/BlobClient.java
@@ -10,6 +10,7 @@ import com.azure.storage.blob.implementation.util.ModelHelper;
 import com.azure.storage.blob.models.AccessTier;
 import com.azure.storage.blob.models.BlobRequestConditions;
 import com.azure.storage.blob.models.BlobHttpHeaders;
+import com.azure.storage.blob.models.BlobStorageException;
 import com.azure.storage.blob.models.ParallelTransferOptions;
 import com.azure.storage.blob.specialized.AppendBlobClient;
 import com.azure.storage.blob.specialized.BlobClientBase;
@@ -167,22 +168,21 @@ public class BlobClient extends BlobClientBase {
     public void uploadWithResponse(InputStream data, long length, ParallelTransferOptions parallelTransferOptions,
         BlobHttpHeaders headers, Map<String, String> metadata, AccessTier tier, BlobRequestConditions requestConditions,
         Duration timeout, Context context) {
-        BlockBlobClient blockBlobClient = this.getBlockBlobClient();
         final ParallelTransferOptions validatedParallelTransferOptions =
             ModelHelper.populateAndApplyDefaults(parallelTransferOptions);
-        if (length < validatedParallelTransferOptions.getMaxSingleUploadSize()) {
-            blockBlobClient.uploadWithResponse(data, length, headers, metadata, tier, null, requestConditions,
-                timeout, context);
-        } else {
-            BlobOutputStream blobOutputStream = BlobOutputStream.blockBlobOutputStream(client,
-                validatedParallelTransferOptions, headers, metadata, tier, requestConditions);
-            try {
-                StorageImplUtils.copyToOutputStream(data, length, blobOutputStream);
-                blobOutputStream.close();
-            } catch (IOException e) {
-                e.printStackTrace();
+        BlobOutputStream blobOutputStream = BlobOutputStream.blockBlobOutputStream(client,
+            validatedParallelTransferOptions, headers, metadata, tier, requestConditions);
+        try {
+            StorageImplUtils.copyToOutputStream(data, length, blobOutputStream);
+            blobOutputStream.close();
+        } catch (IOException e) {
+            if (e.getCause() instanceof BlobStorageException) {
+                throw logger.logExceptionAsError((BlobStorageException) e.getCause());
+            } else {
+                throw logger.logExceptionAsError(new UncheckedIOException(e));
             }
         }
+
     }
 
     /**